feat: refactor entry pages
Simplify entries list. Add an independent entry view page. Simplify edit
entry page.
