+++ lib/widgets/entry_day_cell.dart
@@ -1,7 +1,9 @@
 import 'dart:ui' as ui;
 
 import 'package:daily_you/config_provider.dart';
+import 'package:daily_you/models/entry.dart';
 import 'package:daily_you/models/image.dart';
+import 'package:daily_you/pages/edit_entry_page.dart';
 import 'package:daily_you/providers/entries_provider.dart';
 import 'package:daily_you/providers/entry_images_provider.dart';
 import 'package:daily_you/time_manager.dart';
@@ -10,10 +12,7 @@ import 'package:flutter/material.dart';
 import 'package:daily_you/widgets/mood_icon.dart';
 import 'package:provider/provider.dart';
 import 'package:table_calendar/table_calendar.dart';
-
-import '../models/entry.dart';
-import '../pages/edit_entry_page.dart';
-import '../pages/entry_detail_page.dart';
+import 'package:daily_you/pages/entries_list_page.dart';
 
 class EntryDayCell extends StatelessWidget {
   final DateTime date;
@@ -71,9 +70,9 @@ class EntryDayCell extends StatelessWidget {
           onTap: () async {
             await Navigator.of(context).push(MaterialPageRoute(
               allowSnapshotting: false,
-              builder: (context) => EntryDetailPage(
-                filtered: false,
+              builder: (context) => EntriesListPage(
                 index: entriesProvider.getIndexOfEntry(entry.id!),
+                getEntries: () => entriesProvider.entries,
               ),
             ));
           },
@@ -109,9 +108,9 @@ class EntryDayCell extends StatelessWidget {
           onTap: () async {
             await Navigator.of(context).push(MaterialPageRoute(
               allowSnapshotting: false,
-              builder: (context) => EntryDetailPage(
-                  filtered: false,
-                  index: entriesProvider.getIndexOfEntry(entry.id!)),
+              builder: (context) => EntriesListPage(
+                  index: entriesProvider.getIndexOfEntry(entry.id!),
+                  getEntries: () => entriesProvider.entries),
             ));
           },
         );
